ref_img to vid fix

+++ src/exiv/components/models/wan/main.py
@@ -722,7 +722,7 @@ class Wan21VaceModel(Wan22Model):
         # embeddings
         orig_shape = list(vace_context.shape)
         vace_context = vace_context.movedim(0, 1).reshape([-1] + orig_shape[2:])
-        c = self.vace_patch_embedding(vace_context).to(vace_context.dtype)
+        c = self.vace_patch_embedding(vace_context.float()).to(vace_context.dtype)
         c = c.flatten(2).transpose(1, 2)
         c = list(c.split(orig_shape[0], dim=0))
